.streamlit/requirements.txt
streamlit
boto3
pandas
psycopg2-binary
SQLAlchemy
plotly
scikit-learn
